Move: 폴더 이동에 따른 경로 이동(navbar)
