+++ src/app/(navbar)/my/menu/_component/MenuSection.tsx
@@ -1,4 +1,4 @@
-import DefaultHeader from "@/app/_component/common/DefaultHeader";
+import DefaultHeader from "@/app/_component/common/layout/DefaultHeader";
 import React, { ReactNode } from "react";
 
 interface Props {
@@ -8,7 +8,7 @@ interface Props {
 }
 const MenuSection = ({ text, backUrl, children }: Props) => {
   return (
-    <section>
+    <section className="w-full">
       <DefaultHeader text={text} redirectUrl={backUrl} theme="default" />
       <article className="w-[87.2%] mx-auto mt-8 box-border">
         {children}
