+++ src/app/(navbar)/my/menu/privacy/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const PrivacyPage = () => {
+  return <div>PrivacyPage</div>;
+};
+
+export default PrivacyPage;
