+++ src/app/(navbar)/my/menu/terms/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const TermsPage = () => {
+  return <div>TermsPage</div>;
+};
+
+export default TermsPage;
